+++ src/main/java/demo/codeexample/config/DataInitializer.java
@@ -1,8 +1,8 @@
 package demo.codeexample.config;
 
-import demo.codeexample.user.doman.Role;
-import demo.codeexample.user.doman.User;
-import demo.codeexample.user.doman.UserRepository;
+import demo.codeexample.user.domain.Role;
+import demo.codeexample.user.domain.User;
+import demo.codeexample.user.domain.UserRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
