+++ src/main/java/demo/codeexample/user/domain/UserRepository.java
@@ -1,4 +1,4 @@
-package demo.codeexample.user.doman;
+package demo.codeexample.user.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
