+++ src/main/java/demo/codeexample/user/web/CreateUserRequestDTO.java
@@ -1,6 +1,6 @@
 package demo.codeexample.user.web;
 
-import demo.codeexample.user.doman.Role;
+import demo.codeexample.user.domain.Role;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
