+++ src/main/java/demo/codeexample/user/web/UserController.java
@@ -1,7 +1,7 @@
 package demo.codeexample.user.web;
 
 import demo.codeexample.user.application.UserService;
-import demo.codeexample.user.doman.Role;
+import demo.codeexample.user.domain.Role;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
